Build dynamic connector for reactive mongo and code refactoring
dynamic connector for reative mongo and code refactor(change packages)
