--- dbs/src/main/resources/application.conf
@@ -1,5 +1,0 @@
-mongo {
-  db = "rmongo"
-  collection = "table1"
-  location = "localhost"
-}
