+++ dbs/src/test/scala/com/knoldus/db/ConnectionTest.scala
@@ -1,4 +1,6 @@
-import DBConnection.Connector
+package com.knoldus.db
+
+import com.knoldus.DBConnection.Connector
 import org.scalatest._
 import reactivemongo.api.DefaultDB
 import reactivemongo.bson.BSON
@@ -6,10 +8,8 @@ import reactivemongo.bson.BSON
 class ConnectorTest extends FlatSpec with Connector{
 
     "DB connection table name" should "table1" in{    
-    val res=connector.name
+    val res=connector("localhost","rmongo","table1").name
     val expectedres="table1"
     assert(res===expectedres)
   }
-    
- 
 }
